chore: split logic into multiple files
This only moves code around, no functional changes.
