+++ buf-common.go
@@ -1,0 +1,16 @@
+package libbpfgo
+
+/*
+#cgo LDFLAGS: -lelf -lz
+#include "libbpfgo.h"
+*/
+import "C"
+
+const (
+	// Maximum number of channels (RingBuffers + PerfBuffers) supported
+	maxEventChannels = 512
+)
+
+var (
+	eventChannels = newRWArray(maxEventChannels)
+)
